SC-8898: made it runnable :) Fixed tests.
