+++ src/services/sync/strategies/consumerActions/SchoolAction.js
@@ -1,14 +1,13 @@
-const BaseConsumerStrategie = require('./BaseConsumerStrategie');
+const BaseConsumerAction = require('./BaseConsumerAction');
 // TODO: place from where it is importat must be fixed later
 const { LDAP_SYNC_ACTIONS } = require('../LDAPSyncer');
 const { SchoolRepo } = require('../../repo');
 
 const defaultOptions = {
 	allowedLogKeys: null,
-	SchoolRepo,
 };
 
-class SchoolAction extends BaseConsumerStrategie {
+class SchoolAction extends BaseConsumerAction {
 	constructor(filterActive = true, options = defaultOptions) {
 		super(LDAP_SYNC_ACTIONS.SYNC_SCHOOL, options);
 		this.filterActive = filterActive;
@@ -16,17 +15,14 @@ class SchoolAction extends BaseConsumerStrategie {
 
 	async action(data = {}) {
 		const { school: schoolData = {} } = data;
-		const school = await this.SchoolRepo.findSchoolByLdapIdAndSystem(
-			schoolData.ldapSchoolIdentifier,
-			schoolData.systems
-		);
+		const school = await SchoolRepo.findSchoolByLdapIdAndSystem(schoolData.ldapSchoolIdentifier, schoolData.systems);
 
 		if (school) {
 			if (school.name !== schoolData.name) {
-				await this.SchoolRepo.updateSchoolName(school._id, schoolData.name);
+				await SchoolRepo.updateSchoolName(school._id, schoolData.name);
 			}
 		} else {
-			await this.SchoolRepo.createSchool(schoolData);
+			await SchoolRepo.createSchool(schoolData);
 		}
 	}
 }
