+++ src/services/sync/strategies/consumerActions/index.js
@@ -1,0 +1,9 @@
+const SchoolAction = require('./SchoolAction');
+const UserAction = require('./UserAction');
+const ClassAction = require('./ClassAction');
+
+module.exports = {
+	SchoolAction,
+	UserAction,
+	ClassAction,
+};
