--- src/services/sync/strategies/consumerStategies/index.js
@@ -1,9 +1,0 @@
-const SchoolAction = require('./SchoolAction');
-const UserAction = require('./UserAction');
-const ClassAction = require('./ClassAction');
-
-module.exports = {
-	SchoolAction,
-	UserAction,
-	ClassAction,
-};
